+++ ondam-backend/src/main/java/com/piveguyz/ondambackend/diaryRecord/command/application/controller/DiaryRecordController.java
@@ -17,7 +17,7 @@ public class DiaryRecordController {
     }
 
     @PostMapping("/sendDiary")
-    public ResponseEntity<String> sendDiary(@RequestParam("diaryId") Integer diaryId){
+    public ResponseEntity<String> sendDiary(@RequestParam("diaryId") Long diaryId){
         boolean result = diaryRecordService.sendDiary(diaryId);
         if(result) {
             return ResponseEntity.status(HttpStatus.ACCEPTED).body("일기가 전송되었습니다.");
